Slug generator - return string for book OR section title
Previously, the slug generator returned a tuple with 1-2 items... the book title slug AND the section title slug.

After these changes, it always returns a string and uses the second item only if the complete section title slug can't be built because it's missing the chapter title in the 3rd parameter.