Include ancestor chapter number in url slug
Previously, `generate_slug` only look up one level for the parent
chapter number if the title itself doesn't have a number.  We need it to
look up all the way for the closest ancestor that has a number and use
it in the slug.

Also change the way `generate_slug` looks for chapter number.  It used
to remove the html tags first (e.g. `<span class="os-number">`) and used
a regular expression to look for the chapter number.  For a title like
"21st Century Economic Development", it'll think that "21" is the
chapter number, but it's actually part of the title.