Fix UnicodeEncodeError in slug generation
The existing slug generation code was found to result in an encoding
error with Polish Physics. The bug only occurred in the legacy cnx
baking flow and not in the new COPS baking pipeline, and this is
likely due to the error only occuring when running the slug code with
Python 2 versus Python 3.